Отрисуй меня полностью
