+++ index.html
@@ -23,7 +23,7 @@
     </section>
 
     <!-- Контейнер для изображений от других пользователей -->
-    <section class="pictures  container">
+    <section class="pictures  container js-pictures">
       <h2 class="pictures__title  visually-hidden">Фотографии других пользователей</h2>
 
       <!-- Поле для загрузки нового изображения на сайт -->
