+++ js/utils.js
@@ -1,5 +1,4 @@
 const getRandomPositiveInteger = (a, b) => {
-
   const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
   const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
   const result = Math.random() * (upper - lower + 1) + lower;
